Edited sulfide code for May 2025 Plate 1
I made a lot of changes to the old sulfide code to format it so that it is similar to the new Dionex code. I also added a QAQC_flag column in case there were any checks that did not pass and need to be written to the dataframe. I also edited the code so it tidies the datafiles automatically at the beginning and you do not have to create it manually.